api response 구조 변경

+++ domain/src/main/java/com/moondroid/damoim/domain/model/MoimAddress.kt
@@ -2,6 +2,6 @@ package com.moondroid.damoim.domain.model
 
 data class MoimAddress(
     val address: String,
-    val lat: Long,
-    val lng: Long
+    val lat: Double,
+    val lng: Double
 )
